replace the version that I typed in by hand with the version that swaggerUI generated.
I'm not at all sure this is a good idea, but since it's the first time I've attempted to write
a swagger API description, I thought it best to just go with what the toolkit did.
The largest differences seems pretty cosmetic, stuff like putting double quoting a lot of strings.